QaiLabs/AREA_51/sii_connector/requirements.txt
# Dependencias principales
selenium>=4.10.0
webdriver-manager>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
pandas>=2.0.3
openpyxl>=3.1.2

# Para desarrollo
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Para generación de PDF (opcional)
reportlab>=4.0.4

# Para automatización de navegador (alternativa a Selenium)
playwright>=1.36.0
